task 4

+++ src/service/API.js
@@ -20,6 +20,59 @@ export const allArticlesCount = async () => {
   }
 };
 
+export const createArticle = async (data, token) => {
+  try {
+    const options = {
+      method: 'POST',
+      headers: {
+        accept: 'application/json',
+        'Content-Type': 'application/json;charset=utf-8',
+        Authorization: `Bearer ${token}`,
+      },
+      body: JSON.stringify(data),
+    };
+    const post = await fetch(`${baseURL}/articles`, options);
+    const response = await post.json();
+    return response.article;
+  } catch (error) {
+    throw new Error(error.response);
+  }
+};
+
+export const editArticle = async (data, slug, token) => {
+  try {
+    const options = {
+      method: 'PUT',
+      headers: {
+        accept: 'application/json',
+        'Content-Type': 'application/json;charset=utf-8',
+        Authorization: `Bearer ${token}`,
+      },
+      body: JSON.stringify(data),
+    };
+    const put = await fetch(`${baseURL}/articles/${slug}`, options);
+    const response = await put.json();
+    return response.article;
+  } catch (error) {
+    throw new Error(error.response);
+  }
+};
+
+export const deleteArticle = async (slug) => {
+  try {
+    const token = localStorage.getItem('token');
+    const options = {
+      method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    };
+    await fetch(`${baseURL}/articles/${slug}`, options);
+  } catch (error) {
+    throw new Error(error.response);
+  }
+};
+
 export const registerUser = async (userData) => {
   try {
     const options = {
@@ -31,17 +84,16 @@ export const registerUser = async (userData) => {
       body: JSON.stringify(userData),
     };
     const post = await fetch(`${baseURL}/users`, options);
+    if (post.status === 422) throw new Error('Username or email is already taken');
     const response = await post.json();
-    // console.log('response in API', response.user);
     localStorage.setItem('token', response.user.token);
     return response.user;
   } catch (error) {
-    throw error.response.errors;
+    throw error.response;
   }
 };
 
 export const updateUser = async (userData, token) => {
-  // console.log('data in api before post', userData, 'token:', token);
   try {
     const options = {
       method: 'PUT',
@@ -53,12 +105,10 @@ export const updateUser = async (userData, token) => {
       body: JSON.stringify(userData),
     };
     const post = await fetch(`${baseURL}/user`, options);
-    console.log('post response in api', post);
     if (!post.ok) {
-      throw new Error('try again');
+      throw new Error('failed to fetch');
     }
     const response = await post.json();
-    // console.log('update user in API', response);
     return response.user;
   } catch (error) {
     throw new Error(error.response);
@@ -72,12 +122,10 @@ export const getUser = async (token) => {
         Authorization: `Bearer ${token}`,
       },
     });
-    console.log('get user response', response);
     const data = await response.json();
-    console.log('get user API', data);
     return data;
   } catch (error) {
-    throw error.response.errors;
+    throw error.response;
   }
 };
 
@@ -96,6 +144,40 @@ export const loginUser = async (userData) => {
     localStorage.setItem('token', response.user.token);
     return response.user;
   } catch (error) {
-    throw new Error(error.response.message);
+    throw new Error(error.response);
+  }
+};
+
+export const setLike = async (slug, token) => {
+  try {
+    const options = {
+      method: 'POST',
+      headers: {
+        accept: 'application/json',
+        'Content-Type': 'application/json;charset=utf-8',
+        Authorization: `Bearer ${token}`,
+      },
+    };
+    const post = await fetch(`${baseURL}/articles/${slug}/favorite`, options);
+    const response = await post.json();
+    return response.article;
+  } catch (error) {
+    throw new Error(error.response);
+  }
+};
+
+export const deleteLike = async (slug, token) => {
+  try {
+    const options = {
+      method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    };
+    const onDelete = await fetch(`${baseURL}/articles/${slug}/favorite`, options);
+    const response = await onDelete.json();
+    return response.article;
+  } catch (error) {
+    throw new Error(error.response);
   }
 };
